+++ okta/user.go
@@ -1,6 +1,7 @@
 package okta
 
 import (
+	"encoding/json"
 	"fmt"
 	"log"
 	"net/http"
@@ -56,8 +57,9 @@ var userProfileDataSchema = map[string]*schema.Schema{
 		Computed: true,
 	},
 	"custom_profile_attributes": &schema.Schema{
-		Type:     schema.TypeMap,
-		Computed: true,
+		Type:      schema.TypeString,
+		StateFunc: normalizeDataJSON,
+		Computed:  true,
 	},
 	"department": &schema.Schema{
 		Type:     schema.TypeString,
@@ -187,10 +189,8 @@ func buildUserDataSourceSchema(target map[string]*schema.Schema) map[string]*sch
 }
 
 func assignAdminRolesToUser(u string, r []string, c *okta.Client) error {
-	validRoles := []string{"SUPER_ADMIN", "ORG_ADMIN", "API_ACCESS_MANAGEMENT_ADMIN", "APP_ADMIN", "USER_ADMIN", "MOBILE_ADMIN", "READ_ONLY_ADMIN", "HELP_DESK_ADMIN"}
-
 	for _, role := range r {
-		if contains(validRoles, role) {
+		if contains(validAdminRoles, role) {
 			roleStruct := okta.Role{Type: role}
 			_, _, err := c.User.AddRoleToUser(u, roleStruct)
 
@@ -220,8 +220,13 @@ func assignGroupsToUser(u string, g []string, c *okta.Client) error {
 func populateUserProfile(d *schema.ResourceData) *okta.UserProfile {
 	profile := okta.UserProfile{}
 
-	if _, ok := d.GetOk("custom_profile_attributes"); ok {
-		for k, v := range d.Get("custom_profile_attributes").(map[string]interface{}) {
+	if rawAttrs, ok := d.GetOk("custom_profile_attributes"); ok {
+		var attrs map[string]interface{}
+		str := rawAttrs.(string)
+
+		// We validate the JSON, no need to check error
+		json.Unmarshal([]byte(str), &attrs)
+		for k, v := range attrs {
 			profile[k] = v
 		}
 	}
@@ -398,7 +403,7 @@ func isCustomUserAttr(key string) bool {
 	return !contains(profileKeys, key)
 }
 
-func flattenUser(u *okta.User) map[string]interface{} {
+func flattenUser(u *okta.User, d *schema.ResourceData) (map[string]interface{}, error) {
 	customAttributes := make(map[string]interface{})
 	attrs := map[string]interface{}{}
 
@@ -407,17 +412,37 @@ func flattenUser(u *okta.User) map[string]interface{} {
 			attrKey := camelCaseToUnderscore(k)
 
 			if isCustomUserAttr(attrKey) {
-				// Avoid Terraform blowing up due to wrong type
+				// Supporting any potential type
 				ref := reflect.ValueOf(v)
-				customAttributes[k] = ref.String()
+				switch ref.Kind() {
+				case reflect.String:
+					customAttributes[k] = ref.String()
+				case reflect.Float64:
+					customAttributes[k] = ref.Float()
+				case reflect.Int:
+					customAttributes[k] = ref.Int()
+				case reflect.Bool:
+					customAttributes[k] = ref.Bool()
+				case reflect.Slice:
+					rawArr := v.([]interface{})
+					customAttributes[k] = rawArr
+				case reflect.Map:
+					rawMap := v.(map[string]interface{})
+					customAttributes[k] = rawMap
+				}
 			} else {
 				attrs[attrKey] = v
 			}
 		}
 	}
-	attrs["custom_profile_attributes"] = customAttributes
 
-	return attrs
+	data, err := json.Marshal(customAttributes)
+	if err != nil {
+		return nil, fmt.Errorf("Failed to load custom_attributes to JSON")
+	}
+	attrs["custom_profile_attributes"] = string(data)
+
+	return attrs, nil
 }
 
 // need to remove from all current admin roles and reassign based on terraform configs when a change is detected